move fixed point parsing to fixed point package
